* actorMoveLength is a byte now * camel case * changed some pathfinding functions to use bytes instead of ints * CL_MoveLength now longer uses CL_MoveMode * introduced IS_BMODEL macro

+++ src/client/cl_actor.h
@@ -68,7 +68,7 @@ extern le_t *selActor;
 extern pos3_t truePos;
 extern pos3_t mousePos;
 extern int mousePosTargettingAlign;
-extern int actorMoveLength;
+extern byte actorMoveLength;
 
 extern pos_t *fb_list[MAX_FORBIDDENLIST];
 extern int fb_length;
